+++ model_catalogs/utils.py
@@ -12,7 +12,7 @@ import numpy as np
 import pandas as pd
 import requests
 import yaml
-
+from intake.catalog import Catalog
 from siphon.catalog import TDSCatalog
 
 import model_catalogs as mc
@@ -24,7 +24,7 @@ def astype(value, type_):
     Particularly made to work correctly for returning string, `PosixPath`, or `Timestamp` as list.
     """
     if not isinstance(value, type_):
-        if type_ == list and isinstance(value, (str, pathlib.PurePath, pd.Timestamp)):
+        if type_ == list and isinstance(value, (str, pathlib.PurePath, pd.Timestamp, Catalog)):
             return [value]
         return type_(value)
     return value
@@ -118,7 +118,7 @@ def file2dt(filename):
     return date
 
 
-def get_fresh_parameter(filename):
+def get_fresh_parameter(filename, source):
     """Get freshness parameter, based on the filename.
 
     A freshness parameter is stored in ``__init__`` for required scenarios which is looked up using the
@@ -137,21 +137,42 @@ def get_fresh_parameter(filename):
 
     # a start or end datetime file
     if filename.parent == mc.CACHE_PATH_AVAILABILITY:
-        model_source = filename.name.split("_")[1]
-        if model_source not in mc.FRESH.keys():
-            model_source = "default"
+        # which type are we after
         if "start" in filename.name:
-            mu = mc.FRESH[model_source]["start"]
+            parameter = "start"
         elif "end" in filename.name:
-            mu = mc.FRESH[model_source]["end"]
+            parameter = "end"
         elif "catrefs" in filename.name:
-            mu = mc.FRESH[model_source]["catrefs"]
+            parameter = "catrefs"
+
+        # check for overriding freshness parameter in source metadata
+        if "freshness" in source.metadata and parameter in source.metadata["freshness"]:
+            mu = source.metadata["freshness"][parameter]
+        else:
+            mu = mc.FRESH[parameter]
+        
+        # # model_source = filename.name.split("_")[1]
+        # # if model_source not in mc.FRESH.keys():
+        # #     model_source = "default"
+        # if "start" in filename.name:
+        #     mu = mc.FRESH[model_source]["start"]
+        # elif "end" in filename.name:
+        #     mu = mc.FRESH[model_source]["end"]
+        # elif "catrefs" in filename.name:
+        #     mu = mc.FRESH[model_source]["catrefs"]
     # a file of file locs for aggregation
     elif filename.parent == mc.CACHE_PATH_FILE_LOCS:
-        model_source = filename.name.split("_")[1]
-        if model_source not in mc.FRESH.keys():
-            model_source = "default"
-        mu = mc.FRESH[model_source]["file_locs"]
+        parameter = "file_locs"
+        # check for overriding freshness parameter in source metadata
+        if "freshness" in source.metadata and parameter in source.metadata["freshness"]:
+            mu = source.metadata["freshness"][parameter]
+        else:
+            mu = mc.FRESH[parameter]
+        
+        # model_source = filename.name.split("_")[1]
+        # if model_source not in mc.FRESH.keys():
+        #     model_source = "default"
+        # mu = mc.FRESH[model_source]["file_locs"]
     # a compiled catalog file
     elif filename.parent == mc.CACHE_PATH_COMPILED:
         mu = mc.FRESH["compiled"]
@@ -159,7 +180,7 @@ def get_fresh_parameter(filename):
     return mu
 
 
-def is_fresh(filename):
+def is_fresh(filename, source=None):
     """Check if file called filename is fresh.
 
     If filename doesn't exist, return False.
@@ -168,6 +189,8 @@ def is_fresh(filename):
     ----------
     filename : Path
         Filename to determine freshness
+    source : Intake Source
+        Source from which to check for an overriding freshness parameter. Is not possible for "compiled" catalog files.
 
     Returns
     -------
@@ -179,7 +202,7 @@ def is_fresh(filename):
     try:
         filetime = pd.Timestamp(filename.stat().st_mtime_ns).tz_localize("UTC")
 
-        mu = get_fresh_parameter(filename)
+        mu = get_fresh_parameter(filename, source=source)
 
         return now - filetime < pd.Timedelta(mu)
 
@@ -264,11 +287,13 @@ def find_bbox(ds, dd=None, alpha=None):
 
         # need to calculate concave hull or alphashape of grid
         import alphashape
+        
+        # downsample a bit to save time, still should clearly see shape of domain
+        lon, lat = lon[::dd], lat[::dd]
+        pts = list(zip(lon, lat))
 
         # low res, same as convex hull
-        p0 = alphashape.alphashape(list(zip(lon, lat)), 0.0)
-        # downsample a bit to save time, still should clearly see shape of domain
-        pts = shapely.geometry.MultiPoint(list(zip(lon[::dd], lat[::dd])))
+        p0 = alphashape.alphashape(pts, 0.0)
         p1 = alphashape.alphashape(pts, alpha)
 
     # useful things to look at: p.wkt  #shapely.geometry.mapping(p)
@@ -490,7 +515,7 @@ def find_filelocs(catref, catloc, filetype="fields"):
     return filelocs
 
 
-def calculate_boundaries(file_locs=None, save_files=True, return_boundaries=False):
+def calculate_boundaries(cats, save_files=True, return_boundaries=False):
     """Calculate boundary information for all models.
 
     This loops over all catalog files available in ``mc.CAT_PATH_ORIG``, will try with multiple model_source if necessary (in case servers aren't working) to access the example model output files and calculate the bounding box and numerical domain boundary. The numerical domain boundary is calculated using `alpha_shape` with previously-chosen parameters stored in the original model catalog files. The bounding box and boundary string representation (as WKT) are then saved to files.
@@ -499,7 +524,7 @@ def calculate_boundaries(file_locs=None, save_files=True, return_boundaries=Fals
 
     Parameters
     ----------
-    file_locs : Path, list of Paths, optional
+    cats : Path, list of Paths, optional
         List of Path objects for model catalog files to read from. If not input, will use all catalog files available at ``mc.CAT_PATH_ORIG.glob("*.yaml")``.
     save_files : boolean, optional
         Whether to save files or not. Defaults to True. Saves to ``mc.CAT_PATH_BOUNDARY / cat_loc.name``.
@@ -517,31 +542,26 @@ def calculate_boundaries(file_locs=None, save_files=True, return_boundaries=Fals
 
     >>> mc.calculate_boundaries([mc.CAT_PATH_ORIG / "cbofs.yaml"])
     """
-
-    if file_locs is None:
-        file_locs = mc.CAT_PATH_ORIG.glob("*.yaml")
-    else:
-        file_locs = mc.astype(file_locs, list)
-
+    
     # loop over all orig catalogs
     boundaries = {}
-    for cat_loc in file_locs:
+    for cat in mc.astype(cats, list):
 
-        # open model catalog
-        cat_orig = intake.open_catalog(cat_loc)
+        # # open model catalog
+        # cat_orig = intake.open_catalog(cat)
 
         # loop over available sources and use the first that works
-        for model_source in list(cat_orig):
-            source_orig = cat_orig[model_source]
+        for model_source in list(cat):
+            source_orig = cat[model_source]
             source_transform = mc.transform_source(source_orig)
 
             # need to make catalog to transfer information properly from
             # source_orig to source_transform
             cat_transform = mc.make_catalog(
                 source_transform,
-                full_cat_name=cat_orig.name,  # model name
-                full_cat_description=cat_orig.description,
-                full_cat_metadata=cat_orig.metadata,
+                full_cat_name=cat.name,  # model name
+                full_cat_description=cat.description,
+                full_cat_metadata=cat.metadata,
                 cat_driver=mc.process.DatasetTransform,
                 cat_path=None,
                 save_catalog=False,
@@ -554,8 +574,8 @@ def calculate_boundaries(file_locs=None, save_files=True, return_boundaries=Fals
                 break
 
         # find boundary information for model
-        if "alpha_shape" in cat_orig.metadata:
-            dd, alpha = cat_orig.metadata["alpha_shape"]
+        if "alpha_shape" in cat.metadata:
+            dd, alpha = cat.metadata["alpha_shape"]
         else:
             dd, alpha = None, None
         lonkey, latkey, bbox, wkt = mc.find_bbox(ds, dd=dd, alpha=alpha)
@@ -564,11 +584,11 @@ def calculate_boundaries(file_locs=None, save_files=True, return_boundaries=Fals
 
         # save boundary info to file
         if save_files:
-            with open(mc.FILE_PATH_BOUNDARIES(cat_loc.name), "w") as outfile:
+            with open(mc.FILE_PATH_BOUNDARIES(cat.name.lower()), "w") as outfile:
                 yaml.dump({"bbox": bbox, "wkt": wkt}, outfile, default_flow_style=False)
 
         if return_boundaries:
-            boundaries[cat_loc.stem] = {"bbox": bbox, "wkt": wkt}
+            boundaries[cat.name] = {"bbox": bbox, "wkt": wkt}
 
     if return_boundaries:
         return boundaries